// src/app/models/reportN11.ts
export interface ReportN11 {
  createAt?: any;
  lastModifiedAt?: any;
  createdBy?: any;
  lastModifiedBy?: any;
  id: string;
  provider: string;
  blockStatus: string;
  blockAmount: number;
  numberAmountAll: number;
  numberAmountAssigned: number;
  numberAmountAssignedActive: number;
  numberAmountNotAssigned: number;
  feeAmount: number;
}